fix task strategy snapshot contracts
